from Nazim G
No changes needed to CSV data. The first line of the CSV will have the headers.

When creating the XLSX, you must leave one blank row at the top, and one blank column at the left.

The XlsxWriter module may be used. This is because we want the data to be in an "Excel Table." You may use "Light Green Table Style 21". Picture shows how the spreadsheet should look when the file is opened.

$ python3 script.py input.csv # will generate input.xlsx

A more complicated script that is used for other purposes, including generating an XLSM with (many) sheets and tables is attached in case it helps.

You may find this easy with chatGPT or even devv.ai/en