from chatGPT
You're stuck with the Detail view.. when you run Reliability..

Ideally, we would like to see results in the format

Measure   fresh stress aged (order not important, fresh, aged, stress also ok)

out1 val1 val2 val3

etc

But, this is not possible. However, if Detail Transpose worked, then it is trivial to export CSV and use Excel to get the desired view. But, this currently not possible because the stress test values are not showing