chatGPT - report node count and final line (for sim duration)
functionality : It will find the newest CDS.log.XXX file in my home directory (these files are named CDS.log, CDS.log.2, CDS.log.3, etc) It will grep that CDS.log (newest) for "_axlToolDisplayTestPointOutputLog" and use final line containing this string. From that line, it will get the log_file as string enclosed between double-quotes closes to the end of the line It will grep that log file for ^\s*nodes\s+\d+ (that is, a line that looks like : nodes 14500 ) And print that line. And also print the final line from the file